[src]:branches/VS/obsproc_global
Add new grib fields to GFS and GDAS dump jobs for planned upgrade to global model.
The script will go back as far as 10 days for the high-res sst file, 7 days for the 
high-res ice.  It only goes back 1 day for snow as that was deemed all that is 
considered useful.  Warning msgs are posted if unexpected delays are met.

  * scripts/exglobal_dump.sh.ecf: Add section to pick up new files and post warning 
    messages for any unusual delays in availability.
  * jobs/JGDAS_DUMP: Add variables for directories for new ice and sst files.
  * jobs/JGFS_DUMP:  Add variables for directories for new ice and sst files.
  * Release_Notes.txt:  Update with above info and increased disk usage.